+++ Models/Catalogs/DetalleFichaEntregaRecepcion.cs
@@ -8,7 +8,7 @@ namespace Models.Catalogs
 {
     public class DetalleFichaEntregaRecepcion
     {
-        public int litros { get; set; }
+        public float litros { get; set; }
         public Tanque tanque { get; set; }
         public Pipa pipa { get; set; }
         public FichaEntregaRecepcion ficha { get; set; }
